+++ example/lib/main.dart
@@ -49,8 +49,8 @@ class MyHomePage extends StatelessWidget {
         // onSubmitted: (text) {
         //   searchText.value = text;
         // },
-        animation: (child) => AppBarAnimationSlideLeft(child: child),
-        animationOfSpeechBar: (child) => AppBarAnimationSlideDown(child: child),
+        animation: AppBarAnimationSlideLeft.call,
+        animationOfSpeechBar: AppBarAnimationSlideDown.call,
         appBarBuilder: (context) {
           return AppBar(
             title: Text(title),